move Video to video folder
